requirements.txt
requests>=2.28.0
einops>=0.7.0
gradio
h5py>=3.10.0
huggingface-hub>=0.22.0
imageio
lpips=0.1.4
matplotlib
av>=10.0.0
numpy>=2.0.0
opencv-python
packaging>=24.2
pandas
pillow-simd==9.5.0.post2
protobuf>=4.25.3
scipy
tables
tabulate
termcolor
timm
tqdm
trimesh
triton>=2.4.0
torch==2.9.1+cu130
torchvision==0.24.1+cu130
torchaudio==2.9.1+cu130
wandb
